Replaced char, short, and long with fixed-width alternatives defined in cstdint. This required enabling C++0x, but I think the portability gains are worth it.

+++ transposition.cpp
@@ -9,6 +9,7 @@
 //
 // You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+#include <cstdint>
 #include <iostream>
 #include "evaluate.h"
 #include "expectiminimax.h"
@@ -22,13 +23,13 @@ namespace technicalmachine {
 
 class Hash {
 public:
-	unsigned long ai;
-	unsigned long foe;
-	unsigned long weather;
-	long score;
+	uint64_t ai;
+	uint64_t foe;
+	uint64_t weather;
+	int64_t score;
 	int depth;
 	Hash ();
-	Hash (unsigned long ai_hash, unsigned long foe_hash, unsigned long weather_hash, int depth_current);
+	Hash (uint64_t ai_hash, uint64_t foe_hash, uint64_t weather_hash, int depth_current);
 	bool operator== (Hash const &other) const;
 	bool operator!= (Hash const &other) const;
 };
@@ -37,7 +38,7 @@ Hash::Hash ():
 	depth (0) {
 }
 
-Hash::Hash (unsigned long ai_hash, unsigned long foe_hash, unsigned long weather_hash, int depth_current):
+Hash::Hash (uint64_t ai_hash, uint64_t foe_hash, uint64_t weather_hash, int depth_current):
 	ai (ai_hash),
 	foe (foe_hash),
 	weather (weather_hash),
@@ -52,16 +53,16 @@ bool Hash::operator!= (Hash const &other) const {
 	return !(*this == other);
 }
 
-long transposition (Team &ai, Team &foe, Weather const &weather, int const &depth, score_variables const &sv) {
-	long score;
+int64_t transposition (Team &ai, Team &foe, Weather const &weather, int const &depth, score_variables const &sv) {
+	int64_t score;
 	if (depth == 0)
 		score = evaluate (ai, foe, weather, sv);
 	else {
-		// First, I hash both teams and the weather. These are the long-form hashes that are each the size of an unsigned long, and should be unique within a game. I then take those hashes and divide them modulo their relevant dimension (used to determine the size of the transposition table). These short-form hashes are used as the array index. The shortened hash combinations are used to look up the full hashes, score, and depth.
+		// First, I hash both teams and the weather. These are the long-form hashes that are each the size of a uint64_t, and should be unique within a game. I then take those hashes and divide them modulo their relevant dimension (used to determine the size of the transposition table). These short-form hashes are used as the array index. The shortened hash combinations are used to look up the full hashes, score, and depth.
 		Hash hash (ai.hash(), foe.hash(), weather.hash(), depth);
-		static unsigned long const ai_dimension = 256;
-		static unsigned long const foe_dimension = 256;
-		static unsigned long const weather_dimension = 32;
+		static unsigned const ai_dimension = 256;
+		static unsigned const foe_dimension = 256;
+		static unsigned const weather_dimension = 32;
 		static Hash table [ai_dimension][foe_dimension][weather_dimension] = {};
 		Hash * const transpose = &table [hash.ai % ai_dimension] [hash.foe % foe_dimension] [hash.weather % weather_dimension];
 		// If I can find the current state in my transposition table at a depth of at least the current depth, set the score to the stored score.
